# shared/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Global College LMS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /courses:
    get:
      operationId: listCourses
      tags: [courses]
      summary: List all courses
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: featured
          in: query
          schema:
            type: boolean
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of courses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Course"
    post:
      operationId: createCourse
      tags: [courses]
      summary: Create a course (admin/teacher)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCourseBody"
      responses:
        "201":
          description: Course created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Course"
  /courses/{id}:
    get:
      operationId: getCourse
      tags: [courses]
      summary: Get course by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Course details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CourseDetail"
    put:
      operationId: updateCourse
      tags: [courses]
      summary: Update course
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCourseBody"
      responses:
        "200":
          description: Updated course
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Course"
    delete:
      operationId: deleteCourse
      tags: [courses]
      summary: Delete course
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /lessons:
    get:
      operationId: listLessons
      tags: [lessons]
      summary: List lessons for a course
      parameters:
        - name: courseId
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of lessons
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Lesson"
    post:
      operationId: createLesson
      tags: [lessons]
      summary: Create a lesson
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLessonBody"
      responses:
        "201":
          description: Lesson created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lesson"
...
  /lessons/{id}:
    get:
      operationId: getLesson
      tags: [lessons]
      summary: Get lesson by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Lesson
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lesson"
    put:
      operationId: updateLesson
      tags: [lessons]
      summary: Update lesson
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLessonBody"
      responses:
        "200":
          description: Updated lesson
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lesson"
    delete:
      operationId: deleteLesson
      tags: [lessons]
      summary: Delete lesson
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /lessons/{id}/progress:
    post:
      operationId: updateLessonProgress
      tags: [lessons]
      summary: Mark lesson complete/progress
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LessonProgressBody"
      responses:
        "200":
          description: Progress updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LessonProgress"
  /enrollments:
    get:
      operationId: listEnrollments
      tags: [enrollments]
      summary: List enrollments
      parameters:
        - name: userId
          in: query
          schema:
            type: integer
        - name: courseId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of enrollments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Enrollment"
    post:
      operationId: createEnrollment
      tags: [enrollments]
      summary: Enroll student in course
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEnrollmentBody"
      responses:
        "201":
          description: Enrolled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Enrollment"
  /enrollments/{id}:
    delete:
      operationId: deleteEnrollment
      tags: [enrollments]
      summary: Remove enrollment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Removed
...
components:
  schemas:
...
    Course:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        category:
          type: string
        duration:
          type: string
        fee:
          type: number
        thumbnail:
          type: string
        teacherId:
          type: integer
        teacherName:
          type: string
        isFeatured:
          type: boolean
        isFree:
          type: boolean
        enrollmentCount:
          type: integer
        lessonCount:
          type: integer
        createdAt:
          type: string
      required: [id, title, description, category, duration, fee, isFeatured, isFree, enrollmentCount, lessonCount, createdAt]
    CourseDetail:
      allOf:
        - $ref: "#/components/schemas/Course"
        - type: object
          properties:
            syllabus:
              type: string
            lessons:
              type: array
              items:
                $ref: "#/components/schemas/Lesson"
            assignments:
              type: array
              items:
                $ref: "#/components/schemas/Assignment"
    CreateCourseBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        category:
          type: string
        duration:
          type: string
        fee:
          type: number
        thumbnail:
          type: string
        syllabus:
          type: string
        teacherId:
          type: integer
        isFeatured:
          type: boolean
        isFree:
          type: boolean
      required: [title, description, category, duration, fee]
...
    Lesson:
      type: object
      properties:
        id:
          type: integer
        courseId:
          type: integer
        title:
          type: string
        description:
          type: string
        videoUrl:
          type: string
        pdfUrl:
          type: string
        duration:
          type: integer
        orderIndex:
          type: integer
        isCompleted:
          type: boolean
      required: [id, courseId, title, orderIndex]
    CreateLessonBody:
      type: object
      properties:
        courseId:
          type: integer
        title:
          type: string
        description:
          type: string
        videoUrl:
          type: string
        pdfUrl:
          type: string
        duration:
          type: integer
        orderIndex:
          type: integer
      required: [courseId, title, orderIndex]
    LessonProgressBody:
      type: object
      properties:
        isCompleted:
          type: boolean
        watchedSeconds:
          type: integer
      required: [isCompleted]
    LessonProgress:
      type: object
      properties:
        lessonId:
          type: integer
        userId:
          type: integer
        isCompleted:
          type: boolean
        watchedSeconds:
          type: integer
      required: [lessonId, userId, isCompleted]
    Enrollment:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        courseId:
          type: integer
        courseName:
          type: string
        userName:
          type: string
        progress:
          type: number
        status:
          type: string
          enum: [active, completed, suspended]
        enrolledAt:
          type: string
      required: [id, userId, courseId, progress, status, enrolledAt]
    CreateEnrollmentBody:
      type: object
      properties:
        userId:
          type: integer
        courseId:
          type: integer
      required: [userId, courseId]
    Assignment:
      type: object
      properties:
        id:
          type: integer
        courseId:
          type: integer
        title:
          type: string
        description:
          type: string
        dueDate:
          type: string
        totalMarks:
          type: integer
        createdAt:
          type: string
      required: [id, courseId, title, totalMarks, createdAt]
...